WIP changes to introduce Guids for placed objects
The aim of this was to make it easy to identify objects that have been placed through the use of Guids rather than by comparing the objects themselves.
This was supposed to make things easier as translating objects made it hard to find the correct object to change in the list (before and after). 
However, to make the code not crazy it meant extracting the placedObejcts list into a separate class that could handle the Guids well. This then lent itself to further encapsulation the more I went on. In the end I decided to cut it short as I don't believe such large refactorings would ever be accepted.